+++ redasm/disassembler/disassemblerapi.h
@@ -35,7 +35,7 @@ class DisassemblerAPI
         virtual std::string readString(address_t address) const = 0;
         virtual std::string readWString(const SymbolPtr& symbol) const = 0;
         virtual std::string readWString(address_t address) const = 0;
-        virtual std::string readHex(address_t address, u32 count) const = 0;
+        virtual std::string readHex(address_t address, u64 count) const = 0;
         virtual bool readAddress(address_t address, size_t size, u64 *value) const = 0;
         virtual bool readOffset(offset_t offset, size_t size, u64 *value) const = 0;
         virtual bool getBuffer(address_t address, Buffer& data) const = 0;
