+++ redasm/disassembler/disassemblerbase.h
@@ -26,7 +26,7 @@ class DisassemblerBase: public DisassemblerAPI
         virtual u64 locationIsString(address_t address, bool *wide = NULL) const;
         virtual std::string readString(const SymbolPtr& symbol) const;
         virtual std::string readWString(const SymbolPtr& symbol) const;
-        virtual std::string readHex(address_t address, u32 count) const;
+        virtual std::string readHex(address_t address, u64 count) const;
         virtual SymbolPtr dereferenceSymbol(const SymbolPtr &symbol, u64 *value = NULL);
         virtual bool dereference(address_t address, u64 *value) const;
         virtual bool getBuffer(address_t address, Buffer& data) const;
